Enable score insertion

+++ composeApp/src/commonMain/kotlin/com/belascore/game/domain/model/Score.kt
@@ -1,8 +1,8 @@
 package com.belascore.game.domain.model
 
 data class Score(
-    val gameId: Int,
-    val teamId: Int,
+    val gameId: Long,
+    val teamId: Long,
     val roundNumber: Int,
     val score: Int
 )
